pyproject.toml
[tool.poetry]
name = "ftp-miniapp"
version = "0.1.0"
description = ""
authors = ["Станислав <radchenko_s@stilsoft.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.115.8"}
uvicorn = {extras = ["standard"], version = "^0.34.0"}
tortoise-orm = "^0.24.0"
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = "^2.10.6"
python-multipart = "^0.0.20"
httpx = "^0.28.1"
aiogram = "^3.17.0"
schedule = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
